example cleanup
