+++ .travis.yml
@@ -17,4 +17,5 @@ script:
   - cd examples
   - docker-compose up -d
   - mix deps.get
+  - mix format --check-formatted
   - mix test
