+++ examples/apps/ecto_example/README.md
@@ -1,3 +1,5 @@
-# Ecto instrumentation
+# Ecto
 
 An example of an `Ecto` database app instrumented by the New Relic Agent via `telemetry`.
+
+This instrumentation is fully automatic.
