+++ examples/apps/ecto_example/lib/ecto_example/mysql_repo.ex
@@ -1,0 +1,5 @@
+defmodule EctoExample.MySQLRepo do
+  use Ecto.Repo,
+    otp_app: :ecto_example,
+    adapter: Ecto.Adapters.MyXQL
+end
