+++ examples/apps/ecto_example/lib/ecto_example/postgres_repo.ex
@@ -3,9 +3,3 @@ defmodule EctoExample.PostgresRepo do
     otp_app: :ecto_example,
     adapter: Ecto.Adapters.Postgres
 end
-
-defmodule EctoExample.MySQLRepo do
-  use Ecto.Repo,
-    otp_app: :ecto_example,
-    adapter: Ecto.Adapters.MyXQL
-end
