+++ mix.exs
@@ -11,6 +11,7 @@ defmodule NewRelic.Mixfile do
       start_permanent: Mix.env() == :prod,
       name: "New Relic Elixir Agent",
       source_url: "https://github.com/newrelic/elixir_agent",
+      elixirc_paths: elixirc_paths(Mix.env()),
       package: package(),
       deps: deps(),
       docs: docs()
@@ -53,6 +54,9 @@ defmodule NewRelic.Mixfile do
       extras: ["README.md", "CHANGELOG.md"]
     ]
   end
+
+  defp elixirc_paths(:test), do: ["lib", "test/support"]
+  defp elixirc_paths(_), do: ["lib"]
 
   @agent_version File.read!("VERSION") |> String.trim()
   def agent_version, do: @agent_version
